Implented health endpoint
- Implemented health endpoint
- created a controller
- Added routes
- created server.js
- Added packages to package.json for execution
- updated gitignore
